+++ src/lib.rs
@@ -2,10 +2,12 @@
 //!
 //! ## Features
 //!
-//! - Queries
-//! - Mutations
-//! - Http based connection
-//! - Automatic access token refreshing
+//! -   Queries
+//! -   Mutations
+//! -   Subscriptions
+//! -   HTTP or Websocket connection
+//! -   Automatic access token refreshing
+//! -   Reconnect and resubscribe to subscriptions on socket failure
 //!
 //! ## Example
 //!